Fix ParticipantService type errors after Long→Double amount migration
ParticipantResponse.amount changed to Double; ParticipantService
totalAmount changed to double to match Hisab.amount = Double.

+++ chanlo/src/main/java/com/mysteriousmonkeys/chanlo/dto/ParticipantResponse.java
@@ -9,7 +9,7 @@ public record ParticipantResponse(
     int guestId,
     String guestName,
     String guestPhone,
-    Long amount,
+    Double amount,
     PaymentStatus status,
     LocalDateTime joinedAt,
     String fromStop,
